kubelet - add callback to update PodReadyToStartContainers condition after sandbox creation
This is to address the bug report (gh-issue 1334460), where `PodReadyToStartContainers` condition is only set to `True`, after container image pull is completed, which blocks the pod status managaer update in case the image pull takes significant time..

This commit implements a callback mechanism to allow kubelet to update the `PodReadyToStartContainers` pod condition immediately after the pod sandbox + networking + CSI are ready, but before container images are pulled or containers are created. The callback works as following:
* during kubelet initialization, register the new `onPodSandboxReady` callback function with the runtime manager using type assertion (to maintain backward compatibility)
* during (runtime).SyncPod, after sandbox creation and network configuration, invoke the callback
* the callback retrieves current pod status, generates updated API status, and notifies the status manager to sync to the API server

This callback implementation is gated under `PodReadyToStartContainersCondition` feature gate  and fails gracefully it it logs error but continues pod creation to ensure that this fix don't prevent pod startup.